Rename a model library to a more descriptive namespace
