+++ libsysinspect/src/mdescr/mod.rs
