+++ libsysinspect/src/mdescr/mspec.rs
