+++ libsysinspect/src/mdescr/mspecdef.rs
